pyproject.toml
[project]
name = "audio-swedish-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.1.1",
    "google-cloud-secret-manager>=2.24.0",
    "google-cloud-texttospeech>=2.30.0",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "librosa>=0.11.0",
    "openai-whisper>=20250625",
    "pandas>=2.3.2",
    "pip>=25.2",
    "playwright>=1.55.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchcodec>=0.7.0",
    "webdataset>=1.0.2",
]
